Acceso -user-admin

+++ app/Http/Controllers/AccesoriosController.php
@@ -2,13 +2,13 @@
 
 namespace App\Http\Controllers;
 
-use Illuminate\Http\Request;
-use App\Equipo;
 use App\User;
+use Validator;
+use App\Equipo;
 use App\Accesorio;
 use Carbon\Carbon;
+use Illuminate\Http\Request;
 use Illuminate\Validation\Rule;
-use Validator;
 
 
 class AccesoriosController extends Controller
@@ -22,9 +22,20 @@ class AccesoriosController extends Controller
     {
         $user=auth()->user();
 
-        $accesorios=Accesorio::where('usuario_id', $user->id)->get();
+        if ($user->permisos=='lectura') {
+
+            $accesorios=Accesorio::where('usuario_id', $user->id)->get();
     
-        return view('accesorios.index', compact('accesorios'));
+            return view('usuario.accesorios.index', compact('accesorios'));
+
+        }else if ($user->permisos=='escritura') {
+
+            $accesorios=Accesorio::all();
+
+            return view('admin.accesorios.index', compact('accesorios'));
+
+        }
+        
     }
 
     /**
@@ -34,9 +45,18 @@ class AccesoriosController extends Controller
      */
     public function create()
     {
-        $user = User::pluck('name', 'id');
-        $user->prepend(" "," ");
-        return view('accesorios.crearaccesorio', compact('user'));
+        $user=auth()->user();
+
+        if ($user->permisos=='lectura') {
+            
+            return redirect()->route('accesorios.index');
+
+        }elseif ($user->permisos=='escritura') {
+
+            $user = User::pluck('name', 'id');
+            $user->prepend(" "," ");
+            return view('admin.accesorios.crearaccesorio', compact('user'));
+        }    
     }
 
     /**
@@ -47,66 +67,75 @@ class AccesoriosController extends Controller
      */
     public function store(Request $request)
     {
-        $validator = Validator::make($request->all(),[
-            'usuario_id' => 'required',
-            'fecha_entrega' =>'required',
-            'fecha_mantenimiento' => 'required',
-            'propiedad' => 'required',
-            'tipo_accesorio' => 'required',
-            'marca_accesorio' => 'required|max:100',
-            'referencia_accesorio' => 'required|max:100',
-            'serial_accesorio' => 'required|max:100',
-            'fccid_accesorio' => 'max:100',
-            'icid_accesorio' => 'max:100',
-            'incluido' => 'required',
-            'proveedor' => 'required|max:100',
-            'precio' => 'required|max:100',
-        ],[
-            'required' => 'Este campo es requerido',
-            'email' => 'Este campo debe tener formato de correo electrónico',
-            'unique' => 'Este correo debe ser único',
-            'max' => 'Este campo no debe superar :max caracteres',
-            'min' => 'Este campo no debe ser menor de :min caracteres',
-            'numeric' => 'Este campo debe ser numerico',
-            'string' => 'Este campo debe ser solo texto',
-            'url' => 'Este campo debe ser una url',
-        ]);
+        $user=auth()->user();
 
-        if ($validator->fails()) {
-            return redirect('accesorios/create')
-               ->withErrors($validator)
-               ->withInput();
-         }
+        if ($user->permisos=='lectura') {
+            
+             return redirect()->route('accesorios.index');
 
-        $accesorio = new Accesorio();
-        $accesorio->usuario_id=$request->input('usuario_id');
-        $accesorio->fecha_entrega=Carbon::parse($request->input('fecha_entrega'));
-        $accesorio->fecha_mantenimiento=Carbon::parse($request->input('fecha_mantenimiento'));
+        }elseif ($user->permisos=='escritura') {
 
-        $mantenimiento = Carbon::parse($request->input('fecha_mantenimiento'));
-        $hoy = Carbon::parse(now()->toDateString());
-        $diasDiferencia = $hoy->diffInDays($mantenimiento);
+            $validator = Validator::make($request->all(),[
+                'usuario_id' => 'required',
+                'fecha_entrega' =>'required',
+                'fecha_mantenimiento' => 'required',
+                'propiedad' => 'required',
+                'tipo_accesorio' => 'required',
+                'marca_accesorio' => 'required|max:100',
+                'referencia_accesorio' => 'required|max:100',
+                'serial_accesorio' => 'required|max:100',
+                'fccid_accesorio' => 'max:100',
+                'icid_accesorio' => 'max:100',
+                'incluido' => 'required',
+                'proveedor' => 'required|max:100',
+                'precio' => 'required|max:100',
+            ],[
+                'required' => 'Este campo es requerido',
+                'email' => 'Este campo debe tener formato de correo electrónico',
+                'unique' => 'Este correo debe ser único',
+                'max' => 'Este campo no debe superar :max caracteres',
+                'min' => 'Este campo no debe ser menor de :min caracteres',
+                'numeric' => 'Este campo debe ser numerico',
+                'string' => 'Este campo debe ser solo texto',
+                'url' => 'Este campo debe ser una url',
+            ]);
 
-            if ($diasDiferencia>60) {
-                $accesorio->estado_mantenimiento='PENDIENTE';
-            }else{
-                $accesorio->estado_mantenimiento='BIEN';
-            }
+            if ($validator->fails()) {
+                return redirect('accesorios/create')
+                   ->withErrors($validator)
+                   ->withInput();
+             }
 
-        $accesorio->propiedad=$request->input('propiedad');
-        $accesorio->tipo_de_accesorio=$request->input('tipo_accesorio');
-        $accesorio->marca_accesorio=$request->input('marca_accesorio');
-        $accesorio->referencia_accesorio=$request->input('referencia_accesorio');
-        $accesorio->serial_accesorio=$request->input('serial_accesorio');
-        $accesorio->fccid_accesorio=$request->input('fccid_accesorio');
-        $accesorio->icid_accesorio=$request->input('icid_accesorio');
-        $accesorio->incluido=$request->input('incluido');
-        $accesorio->proveedor=$request->input('proveedor');
-        $accesorio->precio=$request->input('precio');
+            $accesorio = new Accesorio();
+            $accesorio->usuario_id=$request->input('usuario_id');
+            $accesorio->fecha_entrega=Carbon::parse($request->input('fecha_entrega'));
+            $accesorio->fecha_mantenimiento=Carbon::parse($request->input('fecha_mantenimiento'));
 
-        $accesorio->save();
-        
-        return redirect()->route('accesorios.index');
+            $mantenimiento = Carbon::parse($request->input('fecha_mantenimiento'));
+            $hoy = Carbon::parse(now()->toDateString());
+            $diasDiferencia = $hoy->diffInDays($mantenimiento);
+
+                if ($diasDiferencia>60) {
+                    $accesorio->estado_mantenimiento='PENDIENTE';
+                }else{
+                    $accesorio->estado_mantenimiento='BIEN';
+                }
+
+            $accesorio->propiedad=$request->input('propiedad');
+            $accesorio->tipo_de_accesorio=$request->input('tipo_accesorio');
+            $accesorio->marca_accesorio=$request->input('marca_accesorio');
+            $accesorio->referencia_accesorio=$request->input('referencia_accesorio');
+            $accesorio->serial_accesorio=$request->input('serial_accesorio');
+            $accesorio->fccid_accesorio=$request->input('fccid_accesorio');
+            $accesorio->icid_accesorio=$request->input('icid_accesorio');
+            $accesorio->incluido=$request->input('incluido');
+            $accesorio->proveedor=$request->input('proveedor');
+            $accesorio->precio=$request->input('precio');
+
+            $accesorio->save();
+            
+            return redirect()->route('accesorios.index');
+        }
     }
 
     /**
@@ -117,8 +146,18 @@ class AccesoriosController extends Controller
      */
     public function show($id)
     {
-        $accesorio= Accesorio::find($id);
-        return view('accesorios.vermas' , compact('accesorio'));
+        $user=auth()->user();
+
+        if ($user->permisos=='lectura') {
+            
+            $accesorio= Accesorio::find($id);
+            return view('usuario.accesorios.vermas' , compact('accesorio'));
+
+        }elseif ($user->permisos=='escritura') {
+
+            $accesorio= Accesorio::find($id);
+            return view('admin.accesorios.vermas' , compact('accesorio'));
+        }
     }
 
     /**
@@ -129,11 +168,19 @@ class AccesoriosController extends Controller
      */
     public function edit($id)
     {
-        $accesorio= Accesorio::find($id);
-        $user = User::pluck('name', 'id');
-        $user->prepend(" "," ");
+        $user=auth()->user();
 
-        return view('accesorios.editar' , compact('accesorio', 'user'));
+        if ($user->permisos=='lectura') { 
+
+             return redirect()->route('accesorios.index');
+
+        }elseif ($user->permisos=='escritura') {
+            $accesorio= Accesorio::find($id);
+            $user = User::pluck('name', 'id');
+            $user->prepend(" "," ");
+
+            return view('admin.accesorios.editar' , compact('accesorio', 'user'));
+        }
     }
 
     /**
@@ -145,67 +192,76 @@ class AccesoriosController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $validator = Validator::make($request->all(),[
-            'usuario_id' => 'required',
-            'fecha_entrega' =>'required',
-            'fecha_mantenimiento' => 'required',
-            'propiedad' => 'required',
-            'tipo_accesorio' => 'required',
-            'marca_accesorio' => 'required|max:100',
-            'referencia_accesorio' => 'required|max:100',
-            'serial_accesorio' => 'required|max:100',
-            'fccid_accesorio' => 'max:100',
-            'icid_accesorio' => 'max:100',
-            'incluido' => 'required',
-            'proveedor' => 'required|max:100',
-            'precio' => 'required|max:100',
-        ],[
-            'required' => 'Este campo es requerido',
-            'email' => 'Este campo debe tener formato de correo electrónico',
-            'unique' => 'Este correo debe ser único',
-            'max' => 'Este campo no debe superar :max caracteres',
-            'min' => 'Este campo no debe ser menor de :min caracteres',
-            'numeric' => 'Este campo debe ser numerico',
-            'string' => 'Este campo debe ser solo texto',
-            'url' => 'Este campo debe ser una url',
-        ]);
+        $user=auth()->user();
 
-            if ($validator->fails()) {
-                return redirect()
-                   ->back()
-                   ->withErrors($validator)
-                   ->withInput();
-             }
+        if ($user->permisos=='lectura') { 
+            
+             return redirect()->route('accesorios.index');
 
-        $accesorio=Accesorio::find($id);
-        $accesorio->usuario_id=$request->input('usuario_id');
-        $accesorio->fecha_entrega=Carbon::parse($request->input('fecha_entrega'));
-        $accesorio->fecha_mantenimiento=Carbon::parse($request->input('fecha_mantenimiento'));
+        }elseif ($user->permisos=='escritura') {
 
-        $mantenimiento = Carbon::parse($request->input('fecha_mantenimiento'));
-        $hoy = Carbon::parse(now()->toDateString());
-        $diasDiferencia = $hoy->diffInDays($mantenimiento);
+            $validator = Validator::make($request->all(),[
+                'usuario_id' => 'required',
+                'fecha_entrega' =>'required',
+                'fecha_mantenimiento' => 'required',
+                'propiedad' => 'required',
+                'tipo_accesorio' => 'required',
+                'marca_accesorio' => 'required|max:100',
+                'referencia_accesorio' => 'required|max:100',
+                'serial_accesorio' => 'required|max:100',
+                'fccid_accesorio' => 'max:100',
+                'icid_accesorio' => 'max:100',
+                'incluido' => 'required',
+                'proveedor' => 'required|max:100',
+                'precio' => 'required|max:100',
+            ],[
+                'required' => 'Este campo es requerido',
+                'email' => 'Este campo debe tener formato de correo electrónico',
+                'unique' => 'Este correo debe ser único',
+                'max' => 'Este campo no debe superar :max caracteres',
+                'min' => 'Este campo no debe ser menor de :min caracteres',
+                'numeric' => 'Este campo debe ser numerico',
+                'string' => 'Este campo debe ser solo texto',
+                'url' => 'Este campo debe ser una url',
+            ]);
 
-            if ($diasDiferencia>60) {
-                $accesorio->estado_mantenimiento='PENDIENTE';
-            }else{
-                $accesorio->estado_mantenimiento='BIEN';
-            }
+                if ($validator->fails()) {
+                    return redirect()
+                       ->back()
+                       ->withErrors($validator)
+                       ->withInput();
+                 }
 
-        $accesorio->propiedad=$request->input('propiedad');
-        $accesorio->tipo_de_accesorio=$request->input('tipo_accesorio');
-        $accesorio->marca_accesorio=$request->input('marca_accesorio');
-        $accesorio->referencia_accesorio=$request->input('referencia_accesorio');
-        $accesorio->serial_accesorio=$request->input('serial_accesorio');
-        $accesorio->fccid_accesorio=$request->input('fccid_accesorio');
-        $accesorio->icid_accesorio=$request->input('icid_accesorio');
-        $accesorio->incluido=$request->input('incluido');
-        $accesorio->proveedor=$request->input('proveedor');
-        $accesorio->precio=$request->input('precio');
+            $accesorio=Accesorio::find($id);
+            $accesorio->usuario_id=$request->input('usuario_id');
+            $accesorio->fecha_entrega=Carbon::parse($request->input('fecha_entrega'));
+            $accesorio->fecha_mantenimiento=Carbon::parse($request->input('fecha_mantenimiento'));
 
-        $accesorio->save();
-        
-        return redirect()->route('accesorios.index');
+            $mantenimiento = Carbon::parse($request->input('fecha_mantenimiento'));
+            $hoy = Carbon::parse(now()->toDateString());
+            $diasDiferencia = $hoy->diffInDays($mantenimiento);
+
+                if ($diasDiferencia>60) {
+                    $accesorio->estado_mantenimiento='PENDIENTE';
+                }else{
+                    $accesorio->estado_mantenimiento='BIEN';
+                }
+
+            $accesorio->propiedad=$request->input('propiedad');
+            $accesorio->tipo_de_accesorio=$request->input('tipo_accesorio');
+            $accesorio->marca_accesorio=$request->input('marca_accesorio');
+            $accesorio->referencia_accesorio=$request->input('referencia_accesorio');
+            $accesorio->serial_accesorio=$request->input('serial_accesorio');
+            $accesorio->fccid_accesorio=$request->input('fccid_accesorio');
+            $accesorio->icid_accesorio=$request->input('icid_accesorio');
+            $accesorio->incluido=$request->input('incluido');
+            $accesorio->proveedor=$request->input('proveedor');
+            $accesorio->precio=$request->input('precio');
+
+            $accesorio->save();
+            
+            return redirect()->route('accesorios.index');
+        }
     }
 
     /**
@@ -216,9 +272,18 @@ class AccesoriosController extends Controller
      */
     public function destroy($id)
     {
-        $accesorio = Accesorio::find($id);
-        $accesorio->delete();
+       $user=auth()->user();
 
-        return redirect()->route('accesorios.index');
+        if ($user->permisos=='lectura') { 
+            
+             return redirect()->route('accesorios.index');
+
+        }elseif ($user->permisos=='escritura') { 
+
+            $accesorio = Accesorio::find($id);
+            $accesorio->delete();
+
+            return redirect()->route('accesorios.index');
+        }
     }
 }
